Added more features and updated files.

+++ src/Controller/IndexController.php
@@ -3,7 +3,7 @@
 
 namespace App\Controller;
 
-
+use App\Entity\LockupFiles;
 use Doctrine\Persistence\ManagerRegistry;
 use App\Entity\Lockups;
 use App\Entity\Users;
@@ -27,6 +27,8 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Validator\Constraints\Length;
 use WDN\Bundle\FrameworkBundle\Controller\BaseController;
+use Symfony\Component\Console\Output\OutputInterface;
+
 
 use App\Service\Auth;
 use App\Service\LockupsGenerator;
@@ -37,8 +39,9 @@ class IndexController extends BaseController
     /**
      * @Route("/", name="homePage", methods={"GET"})
      */
-    public function homePage(ManagerRegistry $doctrine, Auth $auth): Response
+    public function homePage(ManagerRegistry $doctrine, Auth $auth, array $errorMsg = array("title" => ""), array $fields = [], array $lockupStyle = []): Response
     {
+
         $auth->requireAuth();
         $lockups = $doctrine->getRepository(LockupTemplates::class)->findAll();
         $lockups_fields = $doctrine->getRepository(LockupTemplatesFields::class)->findAll();
@@ -49,29 +52,42 @@ class IndexController extends BaseController
         $serializer = new Serializer($normalizers, $encoders);
         $jsonContent = $serializer->serialize($lockups_fields, 'json', [AbstractNormalizer::ATTRIBUTES => ['slug', 'Uppercase', 'Value']]);
 
+
+        $fields = $serializer->serialize($fields, 'json', [AbstractNormalizer::ATTRIBUTES => ['value', 'fields' => ['slug']]]);
+
         return $this->render('base.html.twig', [
             'page_template' => "createLockups.html.twig",
             'page_name' => "CreateLockups",
             'lockups' => $lockups,
             'lockups_fields' => $lockups_fields,
             'json_lockups_fields' => $jsonContent,
-            'categories' => $lockups_categories
+            'categories' => $lockups_categories,
+            'error_msg' => $errorMsg,
+            'fields' => $fields,
+            'lockup_style' => $lockupStyle
+
         ]);
     }
 
     /**
      * @Route("/", name="addLockup", methods={"POST"})
      */
-    public function addLockup(Request $request, ManagerRegistry $doctrine, ValidatorInterface $validator, Auth $auth, LockupsGenerator $lockupsGenerator)
+    public function addLockup(Request $request, ManagerRegistry $doctrine, ValidatorInterface $validator, Auth $auth, LockupsGenerator $lockupsGenerator, int $id = 0)
     {
         $auth->requireAuth();
+        $errorMsg = array(
+            "title" => "",
+            "body" => ""
+        );
         $fields = $doctrine->getRepository(LockupTemplatesFields::class)->findAll();
 
         $entityManager = $doctrine->getManager();
 
         $lockuptemplate = (int)$request->request->get('lockuptemplate');
         $approver = (int)$request->request->get('approver');
-        $lockups = new Lockups();
+        $institution = (string)$request->get('institution');
+        $department = (string)$request->get('department');
+        $lockups = ($id != 0) ? $doctrine->getRepository(Lockups::class)->find($id) : new Lockups;
         $template = $doctrine->getRepository(LockupTemplates::class)->find($lockuptemplate);
         $arr = [];
         $count = 0;
@@ -79,33 +95,122 @@ class IndexController extends BaseController
         $lockups->setApprover($approver);
         $lockups->setTemplate($template);
         $lockups->setStatus(0);
+        $lockups->setDepartment($department);
+        $lockups->setInstitution($institution);
         $lockups->setUser($auth->getUser());
+        $lockups->setCommunicatorStatus(0);
+        $lockups->setCreativeStatus(0);
+        $lockups->setIsGenerated(0);
+        $lockups->setGenerating(0);
         $errors = $validator->validate($lockups);
-        if (count($errors) > 0) {
-            return $this->render('base.html.twig', [
-                'page_template' => "createLockups.html.twig",
-                'page_name' => "CreateLockups"
-            ]);
-        }
+
+        $lockupsStyle = array(
+            'lockup_id' => $lockups->getId(),
+            'style' => $template->getStyle(),
+            'category' => $template->getCategory()->getSlug(),
+            'template_id' => $template->getId(),
+            'institution' => $institution,
+            'department' => $department,
+            'approver' => $approver
+        );
 
         foreach ($fields as $item) {
-            if (($request->request->get($item->getSlug())) != "") {
+            if (($request->request->get($item->getSlug())) != "" && (($request->request->get($item->getSlug())) != "0")) {
                 $arr[$count] = new LockupsFields;
                 $arr[$count]->setLockup($lockups);
                 $arr[$count]->setFields($item);
                 $arr[$count]->setValue($request->request->get($item->getSlug()));
-                $entityManager->persist($arr[$count]);
+                // $entityManager->persist($arr[$count]); don't persist it now as it might casue an error
                 $count++;
             }
         }
+        // error checking
+        if (count($errors) > 0) {
+            $errorMsg['title'] = "Error";
+            $errorMsg['body'] = "You have an error in your submission.";
+            $response = $this->forward('App\Controller\IndexController::homePage', [
+                'errorMsg' => $errorMsg,
+                'fields' => $arr,
+                'lockupStyle' => $lockupsStyle
+            ]);
+            return $response;
+        }
 
-        $lockup_fields = $doctrine->getRepository(LockupsFields::class)->findAll($lockups->getId());
+
+        foreach ($fields as $item) {
+            if (($request->request->get($item->getSlug())) == "") {
+                $errorMsg['title'] = "Empty Field";
+                $errorMsg['body'] = "Please enter the " . $item->getName() . " field.";
+                $response = $this->forward('App\Controller\IndexController::homePage', [
+                    'errorMsg' => $errorMsg,
+                    'fields' => $arr,
+                    'lockupStyle' => $lockupsStyle
+
+                ]);
+                return $response;
+            }
+        }
+
+
+        if (($department == "") && ($institution == "")) {
+            $errorMsg['title'] = "Error!";
+            $errorMsg['body'] = "Please enter your Institution/Department name.";
+            $response = $this->forward('App\Controller\IndexController::homePage', [
+                'errorMsg' => $errorMsg,
+                'fields' => $arr,
+                'lockupStyle' => $lockupsStyle
+
+            ]);
+            return $response;
+        }
+
+        if (($approver == 0)) {
+            $errorMsg['title'] = "Error!";
+            $errorMsg['body'] = "Please select your Communicator Contract.";
+            $response = $this->forward('App\Controller\IndexController::homePage', [
+                'errorMsg' => $errorMsg,
+                'fields' => $arr,
+                'lockupStyle' => $lockupsStyle
+
+            ]);
+            return $response;
+        }
+
+        // end of error checking
+        if ($id != 0) {
+            // delete existing fields
+            $delete_lockup_fields = $doctrine->getRepository(LockupsFields::class)->findBy(['lockup' => $id]);
+            $delete_entityManager = $doctrine->getManager();
+
+            foreach ($delete_lockup_fields as $item) {
+                $delete_entityManager->remove(($item));
+            }
+
+            //delete existing files
+            $delete_lockup_files = $doctrine->getRepository(LockupFiles::class)->findBy(['lockup' => $id]);
+            foreach ($delete_lockup_files as $item) {
+                $delete_entityManager->remove(($item));
+            }
+            $delete_entityManager->flush();
+        }
+        // now we can finally persist the contents of the array now
+        foreach ($arr as $eachField) {
+            $entityManager->persist($eachField);
+        }
+
+
+        $lockupFields = $doctrine->getRepository(LockupsFields::class)->findBy(['lockup' => $id]);
         $entityManager->persist($lockups);
         $entityManager->flush();
         $lockupsGenerator->createPreview($lockups->getId());
-        // save it in the database and redirect to the manage lockups page
+        if ($id != 0) {
+            return $this->redirectToRoute('previewLockups', [
+                'id' => $id
+            ], 302); 
+        }
         return $this->redirectToRoute('manageLockups', [], 302);
     }
+
     /**
      * @Route("/lockups/create", name="createLockups")
      */
@@ -117,6 +222,7 @@ class IndexController extends BaseController
             '<html><body>The create lockups route Lucky number: ' . $number . '</body></html>'
         );
     }
+
     /**
      * @Route("/lockups/manage", name="manageLockups")
      */
@@ -140,17 +246,21 @@ class IndexController extends BaseController
         $auth->requireAuth();
         $id = $request->request->get('id');
         $lockups = $doctrine->getRepository(Lockups::class)->find($id);
-        $lockup_fields = $doctrine->getRepository(LockupsFields::class)->findAll($lockups->getId());
+        $lockup_fields = $doctrine->getRepository(LockupsFields::class)->findBy(['lockup' => $id]);
         $entityManager = $doctrine->getManager();
 
         foreach ($lockup_fields as $item) {
             $entityManager->remove(($item));
         }
-        $entityManager->remove($lockups);
+        if ($lockups != null)
+        {
+            $entityManager->remove($lockups);
+        }
         $entityManager->flush();
 
         return $this->redirectToRoute('manageLockups', [], 302);
     }
+
     /**
      * @Route("/lockups/preview/{id}", name="previewLockups")
      */
@@ -158,12 +268,22 @@ class IndexController extends BaseController
     {
         $auth->requireAuth();
         $lockup = $doctrine->getRepository(Lockups::class)->find($id);
+
+        if ($lockup == null) {
+            $response = $this->forward('App\Controller\IndexController::errorPage', [
+                'errorTitle' => "Not found!",
+                'errorBody' => "The requested lockup could not be found."
+            ]);
+            return $response;
+        }
         return $this->render('base.html.twig', [
             'page_template' => "previewLockups.html.twig",
             'page_name' => "ManageLockups",
-            'SVG' => $lockup
+            'Lockup' => $lockup,
+            'user' => $auth
         ]);
     }
+
     /**
      * @Route("/logout", name="logout")
      */
@@ -172,6 +292,7 @@ class IndexController extends BaseController
         $auth->logout();
         return $this->redirectToRoute('homePage', [], 302);
     }
+
     /**
      * @Route("/lockups/library", name="lockupsLibrary")
      */
@@ -179,6 +300,128 @@ class IndexController extends BaseController
     {
         return $this->render('base.html.twig', [
             'page_name' => "LockupsLibrary"
+        ]);
+    }
+
+    /**
+     * @Route("/lockups/edit/{id}", name="editLockups", methods={"GET"})
+     */
+    public function editLockups(int $id, ManagerRegistry $doctrine, Auth $auth): Response
+    {
+        $auth->requireAuth();
+        $lockup = $doctrine->getRepository(Lockups::class)->find($id);
+        $lockupFields = $doctrine->getRepository(LockupsFields::class)->findBy(['lockup' => $id]);
+        $lockupsStyle = array(
+            'lockup_id' => $lockup->getId(),
+            'style' => $lockup->getTemplate()->getStyle(),
+            'category' => $lockup->getTemplate()->getCategory()->getSlug(),
+            'template_id' => $lockup->getTemplate()->getId(),
+            'institution' => $lockup->getInstitution(),
+            'department' => $lockup->getDepartment(),
+            'approver' => $lockup->getApprover()
+        );
+        $response = $this->forward('App\Controller\IndexController::homePage', [
+            'fields' => $lockupFields,
+            'lockupStyle' => $lockupsStyle
+
+        ]);
+        return $response;
+    }
+
+    /**
+     * @Route("/lockups/edit/{id}", name="editedLockups", methods={"POST"})
+     */
+    public function editedLockups(int $id, ManagerRegistry $doctrine, Auth $auth): Response
+    {
+        $auth->requireAuth();
+        $response = $this->forward('App\Controller\IndexController::addLockup', [
+            'id' => $id
+
+        ]);
+        return $response;
+    }
+
+    /**
+     * @Route("/lockups/download/{id}", name="downloadLockups", methods={"GET"})
+     */
+    public function downloadLockups(int $id, ManagerRegistry $doctrine, Auth $auth): Response
+    {
+        // $auth->requireAuth();
+        $lockup = $doctrine->getRepository(Lockups::class)->find($id);
+
+        return $this->render('base.html.twig', [
+            'page_template' => "downloadLockups.html.twig",
+            'Lockup' => $lockup
+        ]);
+    }
+
+    /**
+     * @Route("/lockups/generate/{id}", name="generateLockups", methods={"GET"})
+     */
+    public function generateLockups(int $id, ManagerRegistry $doctrine, Auth $auth, LockupsGenerator $lockupsGenerator): Response
+    {
+        // $auth->requireAuth();
+
+        $lockupsGenerator->generateLockups($id);
+
+        return $this->redirectToRoute('downloadLockups', [
+            'id' => $id
+        ], 302);
+    }
+
+   /**
+     * @Route("/lockups/actions/", name="lockupsAction", methods={"POST"})
+     */
+    public function lockupsActions(ManagerRegistry $doctrine, Request $request, Auth $auth): RedirectResponse
+    {
+        $auth->requireAuth();
+        $id = $request->request->get('id');
+        $action = $request->request->get('action');
+        $creative_feedback = $request->request->get('creative_feedback');
+        $communicator_feedback = $request->request->get('communicator_feedback');
+
+        $lockups = $doctrine->getRepository(Lockups::class)->find($id);
+
+        if ($action == "approve" && $auth->isCreative()) {
+            $lockups->setCreativeStatus(1);
+            $lockups->setCreativeFeedback($creative_feedback);
+        } else if ($action == "approve" && $auth->isApprover()) {
+            $lockups->setCommunicatorStatus(1);
+            $lockups->setCommunicatorFeedback($creative_feedback);
+        } else if  ($action == "update" && $auth->isCreative()) {
+            $lockups->setCreativeFeedback($creative_feedback);
+        } else if ($action == "update" && $auth->isApprover()) {
+            $lockups->setCommunicatorFeedback($communicator_feedback);
+        }else if ($action == "feedback" && $auth->isApprover()) {
+            $lockups->setCommunicatorFeedback(0);
+            $lockups->setCommunicatorFeedback($communicator_feedback);
+        }else if ($action == "feedback" && $auth->isCreative()) {
+            $lockups->setCreativeStatus(0);
+            $lockups->setCommunicatorFeedback($communicator_feedback);
+        }else if ($action == "deny" && $auth->isApprover()) {
+            $lockups->setCommunicatorFeedback(0);
+            $lockups->setCommunicatorFeedback($communicator_feedback);
+        }else if ($action == "deny"  && $auth->isCreative()) {
+            $lockups->setCreativeStatus(0);
+            $lockups->setCreativeFeedback($creative_feedback);
+        }
+        $entityManager = $doctrine->getManager();
+        $entityManager->persist($lockups);
+        $entityManager->flush();
+
+        return $this->redirectToRoute('previewLockups', [
+            'id' => $id
+        ], 302); 
+    }
+
+    public function errorPage(string $errorTitle = "", string $errorBody = ""): Response
+    {
+        // $auth->requireAuth();
+
+        return $this->render('base.html.twig', [
+            'page_template' => "error.html.twig",
+            'error_title' => (strlen($errorTitle) == 0) ? "Error" : $errorTitle,
+            'error_body' => (strlen($errorBody) == 0) ? "You have encountered an error." : $errorBody,
         ]);
     }
 }
